move monitor to libary crate
