+++ compute_tools/src/lib.rs
@@ -15,4 +15,3 @@ pub mod params;
 pub mod pg_helpers;
 pub mod spec;
 pub mod sync_sk;
-pub mod vm_monitor;
