+++ libs/vm_monitor/src/protocol.rs
