+++ libs/vm_monitor/src/runner.rs
@@ -14,13 +14,13 @@ use tokio::sync::broadcast;
 use tokio::sync::mpsc;
 use tracing::{info, warn};
 
-use crate::vm_monitor::cgroup::{CgroupWatcher, MemoryLimits, Sequenced};
-use crate::vm_monitor::dispatcher::Dispatcher;
-use crate::vm_monitor::filecache::{FileCacheConfig, FileCacheState};
-use crate::vm_monitor::protocol::{
+use crate::cgroup::{CgroupWatcher, MemoryLimits, Sequenced};
+use crate::dispatcher::Dispatcher;
+use crate::filecache::{FileCacheConfig, FileCacheState};
+use crate::protocol::{
     InboundMsg, InboundMsgKind, OutboundMsg, OutboundMsgKind, Resources,
 };
-use crate::vm_monitor::{bytes_to_mebibytes, get_total_system_memory, Args, MiB};
+use crate::{bytes_to_mebibytes, get_total_system_memory, Args, MiB};
 
 /// Central struct that interacts with informant, dispatcher, and cgroup to handle
 /// signals from the informant.
